requirements.txt
Django==4.2.11
djangorestframework
dj-rest-auth
django-allauth